# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Attendance Gateway API",
    "version": "1.4.0",
    "description": "Zero-Trust Cryptographic Attendance Gateway.\n\n**Authentication**\n- Most client endpoints accept a shared client API key: header `X-Api-Key: ag_...` (mint from the Admin console → API Keys).\n- Role-protected endpoints (professor/admin) use a JWT: `Authorization: Bearer <token>` from `/api/v1/admin/login` or `/api/v1/auth/login`.\n- Global security: `ApiKey OR bearer`; public operations (health, admin login) override with `[]`.\n\n**Building a client?** See the **Client Integration Guide** (QR / 3-second metronome protocol, biometric Gate 2 ordering, timestamp sealing, 250ms window): https://git.atmyhome.tech/het/attendance-gateway/src/branch/main/docs/CLIENT_INTEGRATION_GUIDE.md\n\n**Error payloads:** every 4xx/5xx returns the envelope `{success:false,error:{code,message,latency_ms?,details?}}`. Full code table + client UI actions: **Error Dictionary** — https://git.atmyhome.tech/het/attendance-gateway/src/branch/main/docs/ERROR_DICTIONARY.md (schema: `ErrorEnvelope`).\n\n**To get an API key:** sign in to the Admin console (https://admin.atmyhome.tech) → **API Keys** → Generate Key."
  },
  "servers": [
    {
      "url": "https://api.atmyhome.tech"
    }
  ],
...
  "security": [
    {
      "ApiKey": []
    },
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "Shared client API key (mint from the Admin console)."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/student/device/bind": {
      "post": {
        "tags": [
          "Student"
        ],
        "summary": "Bind this hardware device to the student (HMAC minted)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "bound"
          }
        }
      }
    },
    "/api/v1/student/attendance": {
      "get": {
        "tags": [
          "Student"
        ],
        "summary": "Student attendance history + analytics",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "records + summary"
          }
        }
      }
    },
...
    "/api/v1/students/{studentUuid}/attendance": {
      "get": {
        "tags": [
          "Sessions / Professor"
        ],
        "summary": "A student's attendance ledger",
        "parameters": [
          {
            "name": "studentUuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ledger"
          }
        }
      }
    },
...
  }
}